% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALDB.R
\docType{methods}
\name{dbFetch,GDALDBResult-method}
\alias{dbFetch,GDALDBResult-method}
\title{Retrieve records from GDALDB query}
\usage{
\S4method{dbFetch}{GDALDBResult}(res, n = -1, ...)
}
\description{
Retrieve records from GDALDB query
}
