% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDALDB.R
\docType{methods}
\name{dbSendQuery,GDALDBConnection-method}
\alias{dbSendQuery,GDALDBConnection-method}
\title{Send a query to GDALDB}
\usage{
\S4method{dbSendQuery}{GDALDBConnection}(conn, statement, ...)
}
\description{
Send a query to GDALDB
}
\examples{
# This is another good place to put examples
}
